fin du blog

+++ app/controllers/post_controller.ts
@@ -9,6 +9,8 @@ import {Marked} from 'marked'
 import { markedHighlight } from "marked-highlight";
 import hljs from 'highlight.js';
 import { unlink } from 'fs/promises'
+import { alterPost } from '#abilities/main'
+import PostPolicy from '#policies/post_policy'
 
  @inject()
 export default class PostController {
@@ -73,7 +75,7 @@ export default class PostController {
     const marked = new Marked(
       markedHighlight({
         langPrefix: 'hljs language-',
-        highlight(code, lang, info) {
+        highlight(code, lang) {
           const language = hljs.getLanguage(lang) ? lang : 'plaintext';
           return hljs.highlight(code, { language }).value;
         }
@@ -89,21 +91,29 @@ export default class PostController {
   /**
    * Edit individual record
    */
-  async edit({ params,view }: HttpContext) {
+  async edit({ params,view,session,response,bouncer }: HttpContext) {
     const { id } = params
     const post = await Post.findByOrFail('id', id)
+    if(await bouncer.with(PostPolicy).denies('alterPost',post)){
+      session.flash('error', 'vous n\'avez pas les autorisations pour modifier cet article')
+      return response.redirect().back()
+    }
     return view.render('pages/post/edit', { post })
   }
 
   /**
    * Handle form submission for the edit action
    */
-  async update({ params, request ,session,response }: HttpContext) {
+  async update({ params, request ,session,response,bouncer }: HttpContext) {
     const {id} = params
     const {content,thumbnail,title} =  await request.validateUsing(updatePostValidator)
     const post = await Post.findByOrFail('id', id)
     const slug = post.title !== title && stringHelpers.slug(title).toLocaleLowerCase()
 
+    if(await bouncer.with(PostPolicy).denies('alterPost',post)){
+      session.flash('error', 'vous n\'avez pas les autorisations pour modifier cet article')
+      return response.redirect().back()
+    }
     if(thumbnail){
       await unlink(`public/${post.thumbnail}`)
       const filePath = await this.fileUploaderService.upload(thumbnail,'','posts')
@@ -120,9 +130,13 @@ export default class PostController {
   /**
    * Delete record
    */
-  async destroy({ params ,session,response}: HttpContext) {
+  async destroy({ params ,session,response,bouncer}: HttpContext) {
     const { id } = params
     const post = await Post.findByOrFail('id', id)
+    if(await bouncer.with(PostPolicy).denies('alterPost',post)){
+      session.flash('error', 'vous n\'avez pas les autorisations pour modifier cet article')
+      return response.redirect().back()
+    }
     await unlink(`public/${post.thumbnail}`)
     await post.delete()
     session.flash('success', 'votre article a bien été supprimé')
